test: remove dead stubs, duplicate tests, and misnamed helper files
Dead stub files (comment-only, no tests):
- test_register_loader.py, test_modbus_helpers.py
- test_coordinator_update_cycle.py, test_coordinator_coverage.py
- test_register_loader_filters.py

Complete duplicates (every test already exists in a larger file):
- test_coordinator_error_paths.py (subsumed by test_coordinator_retry_reconnect.py)
- test_modbus_transport.py (subsumed by test_transport_package.py)

Near-complete duplicate:
- test_coordinator_read_cycle.py: 4 of 5 tests duplicated in
  test_coordinator_update_errors.py; unique test
  (test_process_register_value_sensor_unavailable_temperature) moved there

Empty placeholders (docstring only, no tests):
- test_modbus_helpers_retry.py, test_modbus_helpers_request_args.py,
  test_modbus_helpers_errors.py

Helper file renamed:
- test_scanner_coverage.py → helpers_scanner.py (was pure helpers, no tests;
  callers test_scanner_safe_scan.py and test_scanner_full_scan.py updated)

conftest.py: remove three duplicate attribute assignments in mock_coordinator.

No runtime code changed. 242 → 231 test files; test function count unchanged.
