+++ tests/helpers_scanner.py
