--- tests/test_coordinator_coverage.py
@@ -1,1 +1,0 @@
-"""Coordinator coverage tests migrated into focused coordinator test modules."""
