--- tests/test_coordinator_update_cycle.py
@@ -1,1 +1,0 @@
-"""Coordinator update-cycle split complete; targeted tests moved into focused modules."""
