--- tests/test_modbus_helpers.py
@@ -1,1 +1,0 @@
-"""Legacy file retained after split; tests moved to focused modules."""
