--- tests/test_modbus_helpers_errors.py
@@ -1,1 +1,0 @@
-"""Additional modbus helper error-focused tests."""
