--- tests/test_modbus_helpers_request_args.py
@@ -1,1 +1,0 @@
-"""Additional modbus helper request-args tests."""
