--- tests/test_modbus_helpers_retry.py
@@ -1,1 +1,0 @@
-"""Additional modbus helper retry-focused tests."""
